requirements.txt
pydantic
kubernetes
kafka-python